agent/requirements.txt
websockets==12.0
httpx==0.27.0
mss==9.0.1
Pillow==10.3.0
pynput==1.7.6
pyperclip==1.9.0
python-dotenv==1.0.1
# Optional but recommended on real desktop agents — used by agent.publisher_daily
# to publish the screen as a Daily.co camera track. Native binary; install may
# require build tooling on some Linux distros.
daily-python==0.27.0
